handle symfony4 signature
